Change params.yaml format and update ipam
