+++ demo/copy_to_emulated_env/project/playbooks/pni_addlink/params.yaml
@@ -1,9 +1,7 @@
 ---
-source_as: 65550
-dest_as: 65520
-subnet: 169.254.0.0/16
-# subnet: 100.0.0.0/8
-preferred_node: NONE
-# preferred_node: edge-tk01
-redundant_node: NONE
-# redundant_node: edge-tk03
+source_as:
+  asn: 65550
+  subnet: 169.254.0.0/23
+dest_as:
+  asn: 65520
+  subnet: 169.254.2.0/23
